Try some power savings
Shut down the XTENSA CPU core when there is
nothing to be done, and wait for an interrupt.
Although there is no interrupt code here, the
networking code seems to get them when packets
come in, beacons go out, and the OS timer
fire some too, so we never seem to wait too long.

This patch should probably be tested somewhat.

And I don't really have a good way to test power
consumption. It's entirely possible that
beacon frame transmissions draws way more power
than the CPU, in which case this patch is moot.

Interestingly, this patch should improve upon,
but also negate most of the effect of, moving
from 160MHz to 80MHz. Since the CPU will mostly
be idle, little time will be actually running.